Add files via upload
Video games are a topic that people are always happy to discuss. They have transformed from being a medium of entertainment in the times of Tetris to a social phenomenon they are today. Usually, the reasons why games become popular are ambiguous and varied. Still, some factors can be understood through a thorough analysis of dataset "video games sales," to provide some insights on what affects the sales of video games, how the game industry evolved over the years, the preference of games in different countries and so on. 

Our team used R and Excel to clean the data. We did statistical hypotheses testing. Then we utilized R for visualizing data and Python for modeling & prediction. The main business issue is to analyze factors that are affecting the sales of video games across different regions and platforms. This will allow the gaming industry to make informed decisions and create targeted marketing campaigns of games that can be popular at a specific time of the year and particular region.